backend/requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2.10
django-cors-headers==3.10.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
